Updates to support Splunk8s playground environment, largely due to tighter permissions:
CRD added out of band, so `make install` and `make uninstall` updated to accomodate

Updated service account role permissions to be a subset of the default permissions
from SDK (understandably very broad because they don't know what your operator will
want to do) and the permissions granted to users for their playground namespaces.

You can now easily deploy the operator and run single instances of enterprise, but
more advanced deployments are currently blocked by the requirement to set DNSConfig
for pod specs. This is only enabled in k8s v1.10+, and playground is currently
running 1.9. The team is planning to upgrade shortly after the holidays.