Added support for new configuration parameters to set cpu and memory requests and limits for both splunk and spark pods. Set sane defaults for these. This was required to get things to work in playground because otherwise the defaults (0.1 cpu in particular) cause Splunk to take forever to start up.
Made image pull policy configurable both via parameters and environment variables.
Using "Always" for Playground and UCP clusters to ensure that latest actually uses
the latest consistently on them.

Updated "make install" so that it downloads a Splunk NFR and creates a ConfigMap for it
which is now used to set the license for sample enterprise and DFS clusters.