Various updates to get DFS working again with splunk/splunk 7.3.0 images
Note that an ansible bug was discovered that causes ansible_post_tasks
to not work for local files. This fix will likely be included in a 7.3.1
release.

Changed default imagePullPolicy to Always

Always create a spark cluster when enableDFS is true -- if sparkWorkers
is zero or undefined, default to 1

Allow dfs to be enabled on single instances -- seems like DFS may not
support this though?