Added support to automatically cleanup PVCs used by SplunkEnterprise clusters when they are deleted. Note that for this to work, you need to add the following to the meta section of your SplunkEnterprise object definition:
  finalizers:
    - enterprise.splunk.com/delete-pvc

The PVCs used by clusters will only be cleaned up if this is present.

Updated the pod naming to always start with the format "splunk-<name>-" to make
it easier to identify which things are managed by splunk operator, and to
differentiate deployments.

Also added support to config section for schedulerName and affinity parameters
and cleaned up a lot of copy pasta code.