Fixed bug where headless services were named "*-headless". Now they are named "*-service"
A (non-headless) service is now created for search head clusters
named "splunk-<ID>-search-head-service"

Eliminated the need for dnsConfig to be set for pods. This parameter
is not supported by Google (GKE). Instead, FQDN are now used for
SHC configuration.

Tested everything on both open source kubeadm cluster (1.15.1) and GKE