Switched splunk image in operator.yaml to default to splunk/splunk:edge-redhat
Switched single instances, deployer, cluster master and license master from
using Deployment to using StatefulSet. General best practices suggest that
Deployment should only be used for things that do not need to persist state,
and StatefulSet should be used by everything else.

This also solves the problems with PVC failures in multi-zone clusters.
While PVCs need to be created individually outside of a Deployment, the
PVC templates are embedded within StatefulSets. This allows the scheduler
to ensure that volumes are provisioned according to the rules set for the
Pods rather than independent of those constraints. Most notably, it will
ensure that PVCs for any given pod are all created within the same zone,
for multi-zone clusters.