feat: Allow configuration of default TLS in Splunk
By default Splunk will enable SSL/TLS for splunkd (8089) and hec (8088). It can be desirable
to offload encryption between pods to a the network layer using CNI and externally using ingress
or istio. The following environment variables can be added to the operator deployment to disable
default SSL/TLS

This PR depends on splunk/docker-splunk#437
ref #185