fix: Standardize port names and correct service entries
This change aligns the container port names with the convention expected by common mesh solutions such as istio and linkerd format "identifier-proto" for example hec-https when tls mode is enabled and hec-http when disabled. The splunk web port has been removed from the indxer service definitions to be consistent with best practices where the UI is disabled on indexers. This change depends on #194

ref #194