App Framework POC for phase 3.5
This code adds download and install capabilities when an app package
is determined to have changed, without the need for any restart.

App packages are downloaded by the operator then pushed to the
cooresponding pod.  Once successfuly they are either a) installed
for local scope, or b) extracted to the bundle directory for cluster
scope.  Once all apps are finished if any were extracted to a
bundle dir, then a bundle push is performed.  All splunk cmds are
handled via the operator and not ansible.

Ansible still handles the initial install of all apps.