CSPL-1326: Update the File permissions for the cluster scoped apps
Splunk volumes are set with the file permissions +660 by default. When an app
package(with their original file permissions) is included as part of a bundle push,
that app is deployed to the cluster members.  However, if there is a pod reset happens,
then all those apps are set with +660 values. Due to this permissions change, whenever
there is a next bundle push, it is  unnecessarily including those apps as part of the bundle
one more time.

Changes are made such that, always set the permissions to +660, before doing a bundle push,
that way the pod reset scenario will not include the apps(with just file permission changes)
that were already part of the last bundle push